pyproject.toml
[tool.poetry]
name = "ice-braker"
version = "0.1.0"
description = ""
authors = ["Carlos <carlos@pythoncheatsheet.org>"]
license = "MIT"
readme = "README.md"
packages = [{include = "ice_braker"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.30.0"
langchain = "^0.0.161"
black = "^23.3.0"
openai = "^0.27.6"
install = "^1.3.5"
google-search-results = "^2.4.2"
tweepy = "^4.14.0"
flask = "^2.3.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
